+++ __init__.py
@@ -47,9 +47,9 @@ from .preferences import StableDiffusionPreferences
 
 requirements_path_items = (
     # Use the old version of requirements-win.txt to fix installation issues with Blender + PyTorch 1.12.1
-    ('requirements-win-torch-1-11-0.txt', 'Linux/Windows (CUDA)', 'Linux or Windows with NVIDIA GPU'),
-    ('stable_diffusion/requirements-mac-MPS-CPU.txt', 'Apple Silicon', 'Apple M1/M2'),
-    ('stable_diffusion/requirements-lin-AMD.txt', 'Linux (AMD)', 'Linux with AMD GPU'),
+    ('requirements-lin-win-colab-CUDA.txt', 'Linux/Windows (CUDA)', 'Linux or Windows with NVIDIA GPU'),
+    ('requirements-mac-MPS-CPU.txt', 'Apple Silicon', 'Apple M1/M2'),
+    ('requirements-lin-AMD.txt', 'Linux (AMD)', 'Linux with AMD GPU'),
 )
 
 def register():
@@ -58,9 +58,8 @@ def register():
         dt_op = getattr(dt_op, name)
     if hasattr(bpy.types, dt_op.idname()): # objects under bpy.ops are created on the fly, have to check that it actually exists a little differently
         raise RuntimeError("Another instance of Dream Textures is already running.")
-    
 
-    bpy.types.Scene.dream_textures_requirements_path = EnumProperty(name="Platform", items=requirements_path_items, description="Specifies which set of dependencies to install", default='stable_diffusion/requirements-mac-MPS-CPU.txt' if sys.platform == 'darwin' else 'requirements-win-torch-1-11-0.txt')
+    bpy.types.Scene.dream_textures_requirements_path = EnumProperty(name="Platform", items=requirements_path_items, description="Specifies which set of dependencies to install", default='requirements-mac-MPS-CPU.txt' if sys.platform == 'darwin' else 'requirements-lin-win-colab-CUDA.txt')
 
     for cls in PREFERENCE_CLASSES:
         bpy.utils.register_class(cls)
