+++ generator_process/__main__.py
@@ -1,0 +1,7 @@
+import sys
+import os
+sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
+
+from generator_process import main
+
+main()
