+++ requirements-lin-AMD.txt
@@ -1,0 +1,9 @@
+-r requirements.txt
+
+# Get hardware-appropriate torch/torchvision 
+--extra-index-url https://download.pytorch.org/whl/rocm5.1.1 --trusted-host https://download.pytorch.org
+torch
+torchvision
+realesrgan==0.2.5.0
+opencolorio
+-e .
