+++ requirements-lin-win-colab-CUDA.txt
@@ -31,4 +31,5 @@ torch==1.11.0
 # Same as numpy - let pip do its thing
 torchvision
 realesrgan==0.2.5.0
+opencolorio
 -e .
