+++ requirements-mac-MPS-CPU.txt
@@ -1,4 +1,4 @@
--r stable_diffusion/requirements.txt
+-r requirements.txt
 
 --pre
 --extra-index-url https://download.pytorch.org/whl/nightly/cpu --trusted-host https://download.pytorch.org
@@ -7,4 +7,5 @@ protobuf==3.19.4
 torch==1.12.1
 torchvision
 realesrgan==0.2.5.0
+opencolorio
 -e .
